requirements.txt
# ----- Core dependencies -----
numpy<2.0
scipy>=1.10.0
scikit-learn>=1.2.0

# ----- spaCy & NLP ecosystem -----
spacy
thinc
scispacy
negspacy

# ----- Retrieval & matching -----
rank-bm25==0.2.2
nmslib-metabrainz==2.1.3

# ----- Embeddings / Transformers -----
sentence-transformers==5.1.2
transformers==4.57.1
torch>=2.1.0  # or latest stable compatible with your GPU/CPU

# ----- Utilities & other libs -----
click==8.3.0
cloudpathlib==0.23.0
fsspec==2025.9.0
huggingface-hub==0.36.0
joblib==1.5.2
pydantic
PyYAML==6.0.3
tqdm==4.67.1
requests==2.32.5
rich==14.2.0

# ----- Your project (editable) -----
-e .  # editable install of your project (chart‑checker)
